requirements.txt
Flask==3.0.2
firebase-admin==6.4.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
python-dotenv==1.0.1
qrcode==7.4.2
